refactor(areal): split areal into core / manipulators / iostream / debug (#1334)
Last type of Phase 2 group 5a (#1444). areal had the group's largest std::cout
population -- 51 sites -- and almost all of them were already correctly guarded.

    core.hpp        73,412 lines / 345 headers / 0 I/O-family
    areal.hpp       82,620 lines / 372 headers / 5 I/O-family  (was 83,655 / 386 / 5)

Zero I/O-family headers is met. Under-45,000 lines is not, and reporting the
number rather than moving the bar: internal/blocktriple/blocktriple.hpp is 69,758
lines on its own (already 0 I/O), so this core is blocktriple plus about 3,650.
Same ceiling dbns hit (#1449); it belongs to the substrate.

The 51 cout sites were not the problem. The include was.

49 of them already sat behind areal's TRACE_CONVERSION switch, which was already
externally settable -- the defect #1387 found elsewhere was not present here.
What kept four I/O-family headers in every areal translation unit was that the
`#include <iostream>` sat OUTSIDE the guard. It now sits inside it, with the code
that needs it. Same move cfloat's trace includes got in #1417/#1418.

The remaining two:

- constexprClassParameters(), a member function with 19 cout statements, is
  declared in the class and defined out-of-line in debug.hpp -- the shape
  blocktriple's got in #1388.
- assign()'s "assign TBD" stub prints with std::printf instead of std::cout: same
  destination, same text, no stream header. stdout ordering with std::cout is
  guaranteed while sync_with_stdio is on, which is the default.

Two structural finds this surfaced, both of which cost the core its zero

1. numeric_limits.hpp included the areal.hpp UMBRELLA. Since the umbrella
   includes the core, that is a cycle -- core -> numeric_limits -> umbrella ->
   manipulators + iostream -- that #pragma once merely masked, and it left
   core.hpp measuring 82,614 lines with all five I/O-family headers, seven lines
   short of the umbrella itself. It now includes areal_impl.hpp, which is what it
   actually needs.

2. native/integers.hpp, which carries <sstream>, was included unconditionally but
   is used ONLY by the trace blocks. It moved inside the guard, along with
   native/ieee754.hpp: the trace calls to_binary(float), which lives in the text
   half of ieee754 that the core's ieee754_core.hpp deliberately omits. Without
   that second include TRACE_CONVERSION=1 stopped compiling -- a regression this
   PR introduced and the trace gate caught.

Also: to_string() and to_binary() are free functions that format through a
stringstream, so they are manipulators.hpp's; operator<< and operator>> keep
their friend DECLARATIONS in the class (operator>> reaches _fraction directly)
and only their definitions move to iostream.hpp, which is what lets the core get
by with <iosfwd>; native/ieee754.hpp -> native/ieee754_core.hpp;
AREAL_ENABLE_LITERALS and AREAL_THROW_ARITHMETIC_EXCEPTION default in
areal_impl.hpp (#1436); manipulators.hpp now names the headers it uses (#1389).

Verification:

- Layer gate: core.hpp alone, and core.hpp + manipulators.hpp, both compile, link
  and run under -Wall -Wpedantic on gcc 13.3 and clang. The second gate
  INSTANTIATES every manipulator rather than merely including the header -- the
  strengthening #1449 forced after range() reached review with an undetected
  dependency on iostream.hpp.
- Trace gate: TRACE_CONVERSION=1 compiles and prints on both compilers. This is
  the gate that caught find 2 above.
- Self-containment: core, manipulators, iostream, debug, numeric_limits, table
  and the umbrella each compile standalone with 0 errors on both compilers.
- 26 targets built and run green on both compilers, including edu_tables_areals
  and the education example, which exercise the text layers.
- Differential against main: five areal configurations swept over their entire
  encoding space (areal<8,2>, <12,4>, <16,5>, <16,8> exhaustively; <20,6> over
  65,536), each rendered by to_binary (both nibble settings), to_string, to_hex
  (three settings), hex_print, pretty_print, color_print, info_print,
  operator<< under four stream states, the native conversions and the predicates;
  plus a dynamic-range round trip, the relocated constexprClassParameters()
  report and the assign() stub. 72,780,424 bytes byte-identical -- old vs new on
  gcc, old vs new on clang, and gcc vs clang.
- Trace text: unlike dbns, main can switch areal's tracing on as shipped, so the
  comparison is direct. 67,398 bytes over 129 conversions across two
  configurations, byte-identical on both compilers.

Found, not fixed: components(areal) does not compile -- `blockbinary<v.fbits> f;`
uses a non-static member in a constant expression and it calls a decode()
overload that does not exist. It fails identically on main; nothing instantiates
it. Same class as type_field(bfloat16) in #1445.
